+++ Sources/KeyboardKit/UIKit/Buttons/KeyboardButtonRow.swift
@@ -15,7 +15,7 @@ import UIKit
 open class KeyboardButtonRow: UIView, KeyboardStackViewComponent {
     
     public convenience init(
-        height: CGFloat = .standardKeyboardRowHeight,
+        height: CGFloat = standardHeight(for: .current, screen: .main),
         actions: KeyboardActionRow,
         alignment: UIStackView.Alignment = .fill,
         distribution: UIStackView.Distribution = .fillEqually,