+++ Sources/KeyboardKit/UIKit/Components/VerticalKeyboardComponent.swift
@@ -23,6 +23,9 @@ public protocol VerticalKeyboardComponent: UIView {
 
 public extension VerticalKeyboardComponent {
     
+    /**
+     Get/set the height of the component's height constraint.
+     */
     var height: CGFloat {
         get { heightConstraint?.constant ?? intrinsicContentSize.height }
         set { setHeight(to: newValue) }